refactor MealRepository: add userId & checking the food belonging to the user

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -4,13 +4,16 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
-import ru.javawebinar.topjava.to.MealWithExceed;
 import ru.javawebinar.topjava.util.MealsUtil;
+import ru.javawebinar.topjava.web.SecurityUtil;
 
-import java.util.Collection;
+import java.util.ArrayList;
+import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
 
@@ -24,38 +27,55 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
 
     @Override
     public Meal save(Meal meal) {
-        Meal result;
+        Meal result = null;
+        int userId = SecurityUtil.authUserId();
         if (meal.isNew()) {
             meal.setId(idCounter.getAndIncrement());
+            meal.setUserId(userId);
             repository.put(meal.getId(), meal);
             result = meal;
+            log.info("Save new meal with id = {}, userId = {}", result.getId(), userId);
         } else {
-            result = repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+            if (meal.getUserId() == userId) {
+                result = repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+                log.info("Update meal with id = {}, userId = {}", result.getId(), userId);
+            }
         }
-        log.info("Save meal with id = {}", result.getId());
         return result;
     }
 
     @Override
-    public void delete(int id) {
-        repository.remove(id);
-        log.info("Delete meal with id = {}", id);
+    public boolean delete(int id) {
+        boolean result = false;
+        Meal meal = repository.getOrDefault(id, null);
+        int userId = SecurityUtil.authUserId();
+        if (meal != null && userId == meal.getUserId()) {
+            repository.remove(id);
+            log.info("Delete meal with id = {}, userId = {}", id, userId);
+            result = true;
+        }
+        return result;
     }
 
     @Override
     public Meal get(int id) {
-        log.info("Get meal with id = {}", id);
+        int userId = SecurityUtil.authUserId();
+        Meal result = repository.getOrDefault(id, null);
+        if (result != null && userId == result.getUserId()) {
+            log.info("Get meal with id = {}, userId = {}", id, userId);
+        }
         return repository.get(id);
     }
 
     @Override
-    public Collection<Meal> getAll() {
-        return repository.values();
-    }
-
-    @Override
-    public Collection<MealWithExceed> getAllWithExceed() {
-        return MealsUtil.getAllWithExceed(repository.values());
+    public List<Meal> getAll() {
+        int userId = SecurityUtil.authUserId();
+        log.info("getAll, userId = {}", userId);
+        return new ArrayList<>(repository.values())
+                .stream()
+                .filter((meal) -> meal.getUserId() == userId)
+                .sorted(Comparator.comparing(Meal::getDateTime).reversed())
+                .collect(Collectors.toList());
     }
 
 }
